third_party/WebKit/LayoutTests/fast/lists/w3-css3-list-styles-deprecated.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
    description("Test deprectaed list style types");

    shouldBeTrue('CSS.supports("list-style-type", "decimal")');

    shouldBeFalse('CSS.supports("list-style-type", "binary")');
    shouldBeFalse('CSS.supports("list-style-type", "lower-hexadecimal")');
    shouldBeFalse('CSS.supports("list-style-type", "octal")');
    shouldBeFalse('CSS.supports("list-style-type", "upper-hexadecimal")');
    shouldBeFalse('CSS.supports("list-style-type", "afar")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-aa-et")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-aa-er")');
    shouldBeFalse('CSS.supports("list-style-type", "amharic")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-am-et")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-gez")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-abegede")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-abegede-gez")');
    shouldBeFalse('CSS.supports("list-style-type", "amharic-abegede")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-abegede-am-et")');
    shouldBeFalse('CSS.supports("list-style-type", "oromo")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-om-et")');
    shouldBeFalse('CSS.supports("list-style-type", "sidama")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-sid-et")');
    shouldBeFalse('CSS.supports("list-style-type", "somali")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-so-et")');
    shouldBeFalse('CSS.supports("list-style-type", "tigre")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-halehame-tig")');
    shouldBeFalse('CSS.supports("list-style-type", "tigrinya-er")');
    shouldBeFalse('CSS.supports("list-style-type", "tigrinya-er-abegede")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-abegede-ti-er")');
    shouldBeFalse('CSS.supports("list-style-type", "tigrinya-et")');
    shouldBeFalse('CSS.supports("list-style-type", "tigrinya-et-abegede")');
    shouldBeFalse('CSS.supports("list-style-type", "ethiopic-abegede-ti-et")');
    shouldBeFalse('CSS.supports("list-style-type", "lower-norwegian")');
    shouldBeFalse('CSS.supports("list-style-type", "upper-greek")');
    shouldBeFalse('CSS.supports("list-style-type", "upper-norwegian")');
    shouldBeFalse('CSS.supports("list-style-type", "asterisks")');
    shouldBeFalse('CSS.supports("list-style-type", "footnotes")');
</script>
</head>
</html>
