third_party/WebKit/LayoutTests/fast/lists/w3-css3-list-styles-deprecated-expected.txt

Test deprectaed list style types

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS CSS.supports("list-style-type", "decimal") is true
PASS CSS.supports("list-style-type", "binary") is false
PASS CSS.supports("list-style-type", "lower-hexadecimal") is false
PASS CSS.supports("list-style-type", "octal") is false
PASS CSS.supports("list-style-type", "upper-hexadecimal") is false
PASS CSS.supports("list-style-type", "afar") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-aa-et") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-aa-er") is false
PASS CSS.supports("list-style-type", "amharic") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-am-et") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-gez") is false
PASS CSS.supports("list-style-type", "ethiopic-abegede") is false
PASS CSS.supports("list-style-type", "ethiopic-abegede-gez") is false
PASS CSS.supports("list-style-type", "amharic-abegede") is false
PASS CSS.supports("list-style-type", "ethiopic-abegede-am-et") is false
PASS CSS.supports("list-style-type", "oromo") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-om-et") is false
PASS CSS.supports("list-style-type", "sidama") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-sid-et") is false
PASS CSS.supports("list-style-type", "somali") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-so-et") is false
PASS CSS.supports("list-style-type", "tigre") is false
PASS CSS.supports("list-style-type", "ethiopic-halehame-tig") is false
PASS CSS.supports("list-style-type", "tigrinya-er") is false
PASS CSS.supports("list-style-type", "tigrinya-er-abegede") is false
PASS CSS.supports("list-style-type", "ethiopic-abegede-ti-er") is false
PASS CSS.supports("list-style-type", "tigrinya-et") is false
PASS CSS.supports("list-style-type", "tigrinya-et-abegede") is false
PASS CSS.supports("list-style-type", "ethiopic-abegede-ti-et") is false
PASS CSS.supports("list-style-type", "lower-norwegian") is false
PASS CSS.supports("list-style-type", "upper-greek") is false
PASS CSS.supports("list-style-type", "upper-norwegian") is false
PASS CSS.supports("list-style-type", "asterisks") is false
PASS CSS.supports("list-style-type", "footnotes") is false
PASS successfullyParsed is true

TEST COMPLETE
